+++ frontend/src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue' // Импорт метода для создания приложения
 import App from './App.vue' // Импорт главного компонента
-import router from './router' // Маршрутизация
+import router from './app/routes/router' // Маршрутизация
 import 'bootstrap/dist/css/bootstrap.css';
 
 const app = createApp(App); // Создание экземпляра приложения
