+++ frontend/src/pages/instrument/addInstrument.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import http from '../../http-common.js';
+import http from '../../shared/api/http-common.js';
 
 export default {
     data() {
