+++ frontend/src/pages/instrument/listInstrument.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import http from "../../http-common";
+import http from "../../shared/api/http-common";
 export default {
   name: "ListInstrument",
   data() {
