+++ frontend/src/pages/part/listParts.vue
@@ -29,7 +29,7 @@
   </template>
   
   <script>
-  import http from "../../http-common";
+  import http from "../../shared/api/http-common";
   export default {
     name: "ListPart",
     data() {
