+++ frontend/src/pages/techcard/addTechcard.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script>
-import http from '../../http-common.js';
+import http from '../../shared/api/http-common.js';
 
 export default {
     data() {
