+++ frontend/src/pages/techcard/editTechcard.vue
@@ -39,7 +39,7 @@
 </template>
 
 <script>
-import http from '../../http-common.js';
+import http from '../../shared/api/http-common.js';
 
 export default {
     data() {
