+++ frontend/src/shared/ui/navbar.vue
